+++ cli/cmd/migrate_test.go
@@ -40,16 +40,13 @@ func TestMigrate(t *testing.T) {
 	}
 	_, filename, _, _ := runtime.Caller(0)
 	currentDir := path.Dir(filename)
-	cqDir := t.TempDir()
-	defer os.RemoveAll(cqDir)
 
 	for _, tc := range configs {
 		t.Run(tc.name, func(t *testing.T) {
-			defer CloseLogFile()
 			testConfig := path.Join(currentDir, "testdata", tc.config)
-			logFileName := path.Join(cqDir, "cloudquery.log")
 			cmd := NewCmdRoot()
-			cmd.SetArgs([]string{"migrate", testConfig, "--cq-dir", cqDir, "--log-file-name", logFileName})
+			baseArgs := testCommandArgs(t)
+			cmd.SetArgs(append([]string{"migrate", testConfig}, baseArgs...))
 			err := cmd.Execute()
 			if tc.err != "" {
 				assert.Contains(t, err.Error(), tc.err)
@@ -58,7 +55,7 @@ func TestMigrate(t *testing.T) {
 			}
 
 			// check that log was written and contains some lines from the plugin
-			b, logFileError := os.ReadFile(path.Join(cqDir, "cloudquery.log"))
+			b, logFileError := os.ReadFile(baseArgs[3])
 			logContent := string(b)
 			require.NoError(t, logFileError, "failed to read cloudquery.log")
 			require.NotEmpty(t, logContent, "cloudquery.log empty; expected some logs")