+++ cli/cmd/test_connection.go
@@ -85,7 +85,6 @@ func newCmdTestConnection() *cobra.Command {
 		Example: testConnectionExample,
 		Args:    cobra.MinimumNArgs(1),
 		RunE:    testConnection,
-		Hidden:  true,
 	}
 	return cmd
 }
@@ -112,7 +111,7 @@ func testConnection(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return fmt.Errorf("failed to get auth token: %w", err)
 	}
-	teamName, err := auth.GetTeamForToken(authToken)
+	teamName, err := auth.GetTeamForToken(ctx, authToken)
 	if err != nil {
 		return fmt.Errorf("failed to get team name: %w", err)
 	}