+++ plugins/source/github/client/client.go
@@ -10,6 +10,8 @@ import (
 	"time"
 
 	"github.com/bradleyfalzon/ghinstallation/v2"
+	"github.com/cloudquery/httpcache"
+	"github.com/cloudquery/httpcache/diskcache"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
 	"github.com/gofri/go-github-ratelimit/github_ratelimit"
 	"github.com/google/go-github/v59/github"
@@ -31,6 +33,8 @@ type Client struct {
 	orgs            []string
 	orgRepositories map[string][]*github.Repository
 	repos           []string
+
+	Spec Spec
 }
 
 func (c *Client) Logger() *zerolog.Logger {
@@ -91,10 +95,15 @@ func New(ctx context.Context, logger zerolog.Logger, spec Spec) (schema.ClientMe
 		if err != nil {
 			return nil, fmt.Errorf("failed to parse InstallationID %v: %w", auth.InstallationID, err)
 		}
+		transport := http.DefaultTransport
+		if spec.LocalCachePath != "" {
+			cache := diskcache.New(spec.LocalCachePath)
+			transport = httpcache.NewTransport(cache)
+		}
 		if auth.PrivateKeyPath != "" {
-			itr, err = ghinstallation.NewKeyFromFile(http.DefaultTransport, appId, installationId, auth.PrivateKeyPath)
+			itr, err = ghinstallation.NewKeyFromFile(transport, appId, installationId, auth.PrivateKeyPath)
 		} else {
-			itr, err = ghinstallation.New(http.DefaultTransport, appId, installationId, []byte(auth.PrivateKey))
+			itr, err = ghinstallation.New(transport, appId, installationId, []byte(auth.PrivateKey))
 		}
 		if err != nil {
 			return nil, fmt.Errorf("failed to create GitHub client for org %v: %w", auth.Org, err)
@@ -111,7 +120,12 @@ func New(ctx context.Context, logger zerolog.Logger, spec Spec) (schema.ClientMe
 
 	var defaultServices GithubServices
 	if spec.AccessToken != "" {
-		httpClient := github.NewClient(nil).WithAuthToken(spec.AccessToken)
+		var cl *http.Client
+		if spec.LocalCachePath != "" {
+			cache := diskcache.New(spec.LocalCachePath)
+			cl = httpcache.NewTransport(cache).Client()
+		}
+		httpClient := github.NewClient(cl).WithAuthToken(spec.AccessToken)
 		ghc, err := githubClientForHTTPClient(httpClient.Client().Transport, logger, spec.EnterpriseSettings)
 		if err != nil {
 			return nil, fmt.Errorf("failed to create GitHub client for access token: %w", err)
@@ -127,6 +141,7 @@ func New(ctx context.Context, logger zerolog.Logger, spec Spec) (schema.ClientMe
 		orgServices: ghServices,
 		orgs:        spec.Orgs,
 		repos:       spec.Repos,
+		Spec:        spec,
 	}
 	c.logger.Info().Msg("Discovering repositories")
 	orgRepositories, err := c.discoverRepositories(ctx, spec.DiscoveryConcurrency, spec.Orgs, spec.Repos, spec.IncludeArchivedRepos)