+++ plugins/source/github/client/spec.go
@@ -3,8 +3,11 @@ package client
 import (
 	_ "embed"
 	"fmt"
+	"os"
+	"time"
 
 	"github.com/invopop/jsonschema"
+	"github.com/tj/go-naturaldate"
 )
 
 // Spec is the (nested) spec used by GitHub Source Plugin
@@ -25,6 +28,23 @@ type Spec struct {
 	DiscoveryConcurrency int `json:"discovery_concurrency,omitempty" jsonschema:"default=1"`
 	// Include archived repositories when discovering repositories.
 	IncludeArchivedRepos bool `json:"include_archived_repos,omitempty"`
+	// Path to a local directory that will hold the cache. If set, the plugin will cache the GitHub API responses in this directory. Defaults to an empty string (no cache)
+	LocalCachePath string `json:"local_cache_path,omitempty"`
+
+	// Table options to set for specific tables.
+	TableOptions TableOptions `json:"table_options,omitempty"`
+}
+
+type TableOptions struct {
+	// Table options for the github_workflow_runs table.
+	WorkflowRuns WorkflowRunsOptions `json:"github_workflow_runs,omitempty"`
+}
+
+type WorkflowRunsOptions struct {
+	// Time to look back for workflow runs in natural date format. Defaults to all workflows.
+	// Examples: "14 days ago", "last month"
+	CreatedSince    string `json:"created_since,omitempty" jsonschema:"example=14 days ago,example=last month"`
+	ParsedTimeSince string `json:"-"`
 }
 
 type EnterpriseSettings struct {
@@ -86,6 +106,22 @@ func (s *Spec) Validate() error {
 		if err := validateRepo(repo); err != nil {
 			return err
 		}
+	}
+	if s.LocalCachePath != "" {
+		fileInfo, err := os.Stat(s.LocalCachePath)
+		if err != nil && !os.IsNotExist(err) {
+			return fmt.Errorf("error accessing local cache path: %w", err)
+		}
+		if fileInfo != nil && !fileInfo.IsDir() {
+			return fmt.Errorf("local cache path is not a directory")
+		}
+	}
+	if s.TableOptions.WorkflowRuns.CreatedSince != "" {
+		parsedTimeSince, err := naturaldate.Parse(s.TableOptions.WorkflowRuns.CreatedSince, time.Now(), naturaldate.WithDirection(naturaldate.Past))
+		if err != nil {
+			return fmt.Errorf("failed to parse created_since: %w", err)
+		}
+		s.TableOptions.WorkflowRuns.ParsedTimeSince = parsedTimeSince.Format(time.RFC3339)
 	}
 	return nil
 }