+++ plugins/source/github/resources/services/teams/teams_test.go
@@ -8,15 +8,14 @@ import (
 	"github.com/cloudquery/plugin-sdk/v4/faker"
 	"github.com/golang/mock/gomock"
 	"github.com/google/go-github/v59/github"
+	"github.com/stretchr/testify/require"
 )
 
 func buildTeams(t *testing.T, ctrl *gomock.Controller) client.GithubServices {
 	mock := mocks.NewMockTeamsService(ctrl)
 
 	var cs github.Team
-	if err := faker.FakeObject(&cs); err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, faker.FakeObject(&cs))
 	someId := int64(5555555)
 	someURL := "https://github.com/orgs/1/teams/test"
 	cs.MembersURL = &someURL
@@ -26,16 +25,12 @@ func buildTeams(t *testing.T, ctrl *gomock.Controller) client.GithubServices {
 		[]*github.Team{&cs}, &github.Response{}, nil)
 
 	var u github.User
-	if err := faker.FakeObject(&u); err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, faker.FakeObject(&u))
 	mock.EXPECT().ListTeamMembersByID(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Return(
 		[]*github.User{&u}, &github.Response{}, nil)
 
 	var r github.Repository
-	if err := faker.FakeObject(&r); err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, faker.FakeObject(&r))
 	r.Parent = &github.Repository{ID: &someId}
 	r.TemplateRepository = &github.Repository{ID: &someId}
 	r.Source = &github.Repository{ID: &someId}
@@ -44,9 +39,8 @@ func buildTeams(t *testing.T, ctrl *gomock.Controller) client.GithubServices {
 		[]*github.Repository{&r}, &github.Response{}, nil)
 
 	var m github.Membership
-	if err := faker.FakeObject(&m); err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, faker.FakeObject(&m))
+
 	mock.EXPECT().GetTeamMembershipBySlug(gomock.Any(), "testorg", *cs.Slug, *u.Login).Return(
 		&m, &github.Response{}, nil)