+++ plugins/destination/motherduck/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## [1.0.7](https://github.com/cloudquery/cloudquery/compare/plugins-destination-motherduck-v1.0.6...plugins-destination-motherduck-v1.0.7) (2024-04-09)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.38.0 ([#17562](https://github.com/cloudquery/cloudquery/issues/17562)) ([fb0d1b4](https://github.com/cloudquery/cloudquery/commit/fb0d1b4eba8e9d0fcfc436cc1881317f178896bc))
+
+## [1.0.6](https://github.com/cloudquery/cloudquery/compare/plugins-destination-motherduck-v1.0.5...plugins-destination-motherduck-v1.0.6) (2024-04-09)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.36.4 ([#17485](https://github.com/cloudquery/cloudquery/issues/17485)) ([f370de4](https://github.com/cloudquery/cloudquery/commit/f370de449e61244398e6f413b973cbfa15c019a6))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.36.5 ([#17526](https://github.com/cloudquery/cloudquery/issues/17526)) ([554c499](https://github.com/cloudquery/cloudquery/commit/554c499eb9bc9f98f6f3dc4be23fd02049f48dcd))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.37.0 ([#17532](https://github.com/cloudquery/cloudquery/issues/17532)) ([8080970](https://github.com/cloudquery/cloudquery/commit/8080970f40d22b6bc9db4c359780c744b476bb02))
+* Don't call `RowsAffected` on query error ([#17568](https://github.com/cloudquery/cloudquery/issues/17568)) ([68cf566](https://github.com/cloudquery/cloudquery/commit/68cf5663def08d461bbd84d775cc630feefb5c9a))
+
+## [1.0.5](https://github.com/cloudquery/cloudquery/compare/plugins-destination-motherduck-v1.0.4...plugins-destination-motherduck-v1.0.5) (2024-04-02)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/codegen to v0.3.13 ([#17444](https://github.com/cloudquery/cloudquery/issues/17444)) ([da276fe](https://github.com/cloudquery/cloudquery/commit/da276fe64c46ec0a5f182c83ebc32a90d55f5d50))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.36.1 ([#17372](https://github.com/cloudquery/cloudquery/issues/17372)) ([aaf6187](https://github.com/cloudquery/cloudquery/commit/aaf61873ae5d2e01ea5f3b8b319e4f79afb7b29c))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.36.2 ([#17450](https://github.com/cloudquery/cloudquery/issues/17450)) ([2947506](https://github.com/cloudquery/cloudquery/commit/294750650269f8191c6dfff060c4d3a546405763))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.36.3 ([#17456](https://github.com/cloudquery/cloudquery/issues/17456)) ([020865a](https://github.com/cloudquery/cloudquery/commit/020865a6fde8c896947a844021f0cd7daeb01b06))
+
 ## [1.0.4](https://github.com/cloudquery/cloudquery/compare/plugins-destination-motherduck-v1.0.3...plugins-destination-motherduck-v1.0.4) (2024-03-26)